pyproject.toml
[tool.poetry]
name = "repl_python3_Polypedilumvanderplanki2"
version = "0.1.0"
description = ""
authors = ["FatimaSabitova <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.2.4"
biopython = "^1.78"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
